backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
python-dotenv==1.0.1
elasticsearch==8.14.0
pydantic==2.9.2
sentence-transformers==3.0.1
Pillow==10.4.0
imagehash==4.3.1
rapidfuzz==3.9.6
orjson==3.10.7
